requirements.txt
flask==3.0.3
flask-cors==4.0.0
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.32.3
gunicorn==21.2.0
